pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "clv-long-term-optimization"
version = "1.0.0"
description = "Production-style CLV forecasting, churn risk modeling, and budget-constrained retention optimization"
readme = "README.md"
license = { text = "Copyright © 2026 Ramesh Shrestha. All rights reserved." }
authors = [{ name = "Ramesh Shrestha", email = "ramesh@example.com" }]
requires-python = ">=3.9"

dependencies = [
    "numpy>=1.24,<2.0",
    "pandas>=2.1,<2.4",
    "scipy>=1.11,<1.14",
    "scikit-learn>=1.3,<1.7",
    "statsmodels>=0.14,<0.15",
    "lifetimes>=0.11,<0.12",
    "pulp>=2.7,<3.4",
    "pyarrow>=14,<21",
    "openpyxl>=3.1,<3.2",
    "matplotlib>=3.7,<3.10",
    "seaborn>=0.12,<0.14",
    "mlflow>=2.9,<3.2",
    "xgboost>=2.0,<3.0",
    "lightgbm>=4.0,<5.0",
    "shap>=0.44,<1.0",
    "joblib>=1.3,<1.6",
    "tqdm>=4.65,<5.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9.0",
    "pytest-cov>=4.0,<6.0",
    "ruff>=0.3,<1.0",
    "jupyterlab>=4.0,<5.0",
    "ipykernel>=6.25,<7.0",
]

[project.scripts]
clv-pipeline = "src.pipelines.weekly_scoring_pipeline:main"
clv-ingest   = "src.ingestion.load_data:main"
clv-clean    = "src.cleaning.clean_transactions:main"
clv-features = "src.features.build_features:main"
clv-model    = "src.modeling.train_clv_models:main"
clv-churn    = "src.modeling.train_churn_risk:main"
clv-optimize = "src.optimization.budget_allocator:main"
clv-evaluate = "src.evaluation.backtesting:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ── Ruff: linting & formatting ─────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line-too-long (handled by formatter)
    "B008",  # do not perform function calls in default arguments
    "N803",  # argument name should be lowercase — X is standard ML convention (scikit-learn API)
    "N806",  # variable in function should be lowercase — X_train, X_test, X_df are ML convention
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # assert is fine in tests

# ── Pytest ──────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.coverage.run]
source  = ["src"]
omit    = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
show_missing = true
fail_under   = 70
